:truck: move: 記事コンポーネントを同一ディレクトリに集約
