--- src/components/molecules/article/index.ts
@@ -1,3 +1,0 @@
-export { default as ArticleList } from "./ArticleList";
-export { ArticleSkeltonList } from "./ArticleSkeltonList";
-export { ArticleTipWithThumb, ArticleTipWithThumbList } from "./ArticleTipList";
