+++ src/components/organisms/article/ArticleSWRContainer/ArticleSkeltonList/index.tsx
