+++ src/components/organisms/article/ArticleTipList/index.spec.tsx
