+++ src/components/organisms/article/ArticleTipList/index.stories.tsx
